requirements.txt
certifi==2025.4.26
charset-normalizer==3.4.2
delta-spark==3.2.0
idna==3.10
importlib_metadata==8.7.0
py4j==0.10.9.7
pyspark==3.5.6
python-dotenv==1.1.1
ratelimit==2.2.1
requests==2.32.3
retrying==1.4.2
urllib3==2.4.0
zipp==3.23.0
